feat: clean button component
