+++ frontend/src/components/common/Button/index.ts
@@ -1,0 +1,2 @@
+export { Button } from './Button';
+export type { StyleTypes, Variants, Props } from './Button.interface';
